Fix captain push -Bc when on tag
When checked out on a tag and pushing all branches, captain failed to
remove `(HEAD detached at v1.0)` from the `git branch` output. It then
tried to tag a docker container with that text of which docker would
(rightly so) error out. The changes in this commit remove this
extraneous text from the docker label list when on a commit not marked by
a branch. In addition, only branches whose heads point to the current HEAD
are considered, as opposed to all branches that contain the HEAD commit.
Tests were added to `git_test.go` on a git repository to verify it
behaved properly whether the current HEAD commit is pointed at by tags or
branches, just tags, or just branches.